Correct handling negated conjectures.
Before, when we had several negated conjectures,
we would succeed only if we refuted *all* of them.
After this change, we succeed if we refute *at least one* of them.
That way, negated conjectures behave more similarly to axioms
if we have more than one of them.

This should close #1.